TransportScheduler v0.1.0 InputParser

Module to store data from input files in the defined station structure. Files stations.txt, schedule.txt, OMT.txt, local_variables.txt store data of transit stations and connections. These are read into appropriate maps for storing in corresponding Station processes.

Uses GenServer.

Functions

get_city_code(pid, city)

Returns station code from map using station city as key, taken from InputParser process when its pid is known.

Parameters

ip_pid

station_name

Return values

Returns {:reply, station_code, state}.

get_local_variables(pid, code)

Return map of local variables from source station having given code, taken from InputParser process when its pid is known.

Parameters

ip_pid

station_code

Return values

Returns {:reply, loc_var_map, state}.

get_other_means(pid, code)

Returns schedule list of other means connections from source station having given code, taken from InputParser process when its pid is known.

Parameters

ip_pid

station_code

Return values

Returns {:reply, OMT_list, state}.

get_schedule(pid, code)

Return schedule list of vehicles from source station having given code, taken from InputParser process when its pid is known.

Parameters

ip_pid

station_code

Return values

Returns {:reply, schedule_list, state}.

get_schedules(pid)

Returns map of source station codes to schedule list of vehicles from that source station, taken from InputParser process when its pid is known.

Parameters

ip_pid

Return values

Returns {:reply, schedules_map, state}.

get_station_map(pid)

Returns map of station cities to station codes, taken from InputParser process when its pid is known.

Parameters

ip_pid

Return values

Returns {:reply, station_map, state}.

get_station_struct(pid, city)

Return Station struct having local variable and schedule information for a given station city, taken from InputParser process when its pid is known.

Parameters

ip_pid

station_name

Return values

Returns station struct of the form %{loc_vars_map, schedule_list, other_means_list, station_code, station_city}.

obtain_loc_var_map()

obtain_other_means()

obtain_schedules()

obtain_stations()

start_link()

Starts a GenServer InputParser process linked to the current process.

This is often used to start the GenServer as part of a supervision tree.

Once the server is started, the init/1 function of the given module is called with args as its arguments to initialize the server.

Parameters

module

args

Return values

If the server is successfully created and initialized, this function returns {:ok, pid}, where pid is the PID of the server. If a process with the specified server name already exists, this function returns {:error, {:already_started, pid}} with the PID of that process.

If the init/1 callback fails with reason, this function returns {:error, reason}. Otherwise, if it returns {:stop, reason} or :ignore, the process is terminated and this function returns {:error, reason} or :ignore, respectively.

stop(pid)

Stops the InputParser process with the given reason.

Parameters

ip_pid

Return values

The terminate/2 callback of the given server will be invoked before exiting. This function returns :ok if the server terminates with the given reason; if it terminates with another reason, the call exits.