TransportScheduler v0.1.0 Mix.Tasks.AsyncBm

Helper module to run asynchronous benchmarking.

Functions

run(arg)

Runs the AsyncBm task with the given args.

Parameters

arg

Return values

If the task was not yet invoked, it runs the task and returns the result. If there is an alias with the same name, the alias will be invoked instead of the original task. If the task or alias were already invoked, it does not run them again and simply aborts with :noop.

setup()

Initialises the network using HHTPoison for http://localhost:8880/api.