TransportScheduler v0.1.0 Mix.Tasks.Profile

Helper module to run profiling code using ExProf.

Functions

do_analyze()

Runs the Profile task with the given args to analyze the enclosed code.

Return values

If the task was not yet invoked, it runs the task and returns the result. If there is an alias with the same name, the alias will be invoked instead of the original task. If the task or alias were already invoked, it does not run them again and simply aborts with :noop.

run(mix_args)

Prints analysis results.