TransportScheduler v0.1.0 NetworkConstructor

Module to create registry process for monitoring all station processes and queries using a Network Constructor. The NC process can be used to add or remove queries from the active query list and check if a query is active. It can also be used to maintain a map of station pids, station codes, and station city names for convenient retrieval.

Uses GenServer.

Functions

add_query(server, query, conn)

Adds a query to active list maintained at NC.

Parameters

nc_pid

query - in the form of a map %{src_station, dst_station, arrival_time, end_time}

conn

Return values

Returns {:ok}.

check_active(server, query)

Checks if a query is currently in active list maintained at NC.

Parameters

nc_pid

query - in the form of a map %{src_station, dst_station, arrival_time, end_time}

Return values

Returns {:ok}.

create(server, name, code)

Starts a new Station process and adds new pid to registry map using supervisor. Supervisor is also used to restart crashed process and update registry map with new pid.

Parameters

nc_pid

station_name

station_code

Return values

Returns {:ok}.

del_query(server, query)

Removes a query from active list maintained at NC.

Parameters

nc_pid

query - in the form of a map %{src_station, dst_station, arrival_time, end_time}

Return values

Returns {:ok}.

lookup_code(server, code)

Returns station pid and name from registry given the station code.

Parameters

nc_pid

station_code

Return values

Returns {:reply, {name, pid}, state}.

lookup_name(server, name)

Returns station pid and code from registry given the station name.

Parameters

nc_pid

station_name

Return values

Returns {:reply, {code, pid}, state}.

put_queries(server, queries)

return_queries(server)

Returns active query list maintained at NC.

Parameters

nc_pid

Return values

Returns {:reply, queries, state}.

send_to_src(src, dest, itinerary)

Sends a query encoded in itinerary from NC to the source station using receive_at_src/3 of Station module. The pid of NC and source station must be known.

Parameters

nc_pid

src_stn_pid

itinerary - in the form of a map %{vehicleID, src_station, dst_station, dept_time, arrival_time, mode_of_transport}

Return values

Returns {:ok}

start_link(name)

Starts a GenServer NetworkConstructor process linked to the current process.

This is often used to start the GenServer as part of a supervision tree.

Once the server is started, the init/1 function of the given module is called with args as its arguments to initialize the server.

Parameters

module

args

options:

:name - used for name registration

:timeout - if present, the server is allowed to spend the given amount of milliseconds initializing or it will be terminated and the start function will return {:error, :timeout}

:debug - if present, the corresponding function in the :sys module is invoked

:spawn_opt - if present, its value is passed as options to the underlying process

Return values

If the server is successfully created and initialized, this function returns {:ok, pid}, where pid is the PID of the server. If a process with the specified server name already exists, this function returns {:error, {:already_started, pid}} with the PID of that process.

If the init/1 callback fails with reason, this function returns {:error, reason}. Otherwise, if it returns {:stop, reason} or :ignore, the process is terminated and this function returns {:error, reason} or :ignore, respectively.

stop(server)

Stops the NC process with the given reason.

Parameters

nc_pid

Return values

The terminate/2 callback of the given server will be invoked before exiting. This function returns :ok if the server terminates with the given reason; if it terminates with another reason, the call exits.