TransportScheduler v0.1.0 StationFunctions

Module listing possible functions to use to compute congestion_delay value based on delay and some factor for multiplication. Based on the choice of function when Station local variables are being updated, a different mathematical formula gives the final congestion delay value from the original base delay value and the factor value.

Functions

compute_congestion_delay1(delay, factor)

Congestion computation function 1. Computes congestion_delay=delay*factor.

Parameters

delay

factor

Return values

Returns congestion_delay value.

compute_congestion_delay2(delay, factor)

Congestion computation function 2. Computes congestion_delay=delay*factor+0.2.

Parameters

delay

factor

Return values

Returns congestion_delay value.

compute_congestion_delay3(delay, factor)

Congestion computation function 3. Computes congestion_delay=delayfactorfactor.

Parameters

delay

factor

Return values

Returns congestion_delay value.

func(choose_fn)

Chooses congestion computation function

Parameters

choose_fn

Return values

Returns congestion_delay value based on function chosen to compute it.