TransportScheduler v0.1.0 UQCFSM

Module to filter results of the search query in UQC based on user preferences.

Uses GenStateMachine.

Functions

check_bus(list)

Checks if a particular itinerary has only bus as a mode of transport.

check_flight(list)

Checks if a particular itinerary has only flight as a mode of transport.

check_train(list)

#Check if a particular itinerary has only train as a mode of transport

find_transport(itinerary, preference)

Takes in the list of itineraries and preference of the user.

print_pref(list)

Prints the required lists.

print_pref(atom, pref)

start_link()

Starts a GenStateMachine UQCFSM process linked to the current process.

This is often used to start the GenStateMachine as part of a supervision tree.

Once the server is started, the init/1 function of the given module is called with args as its arguments to initialize the server.

Parameters

module

args

Return values

If the server is successfully created and initialized, this function returns {:ok, pid}, where pid is the PID of the server. If a process with the specified server name already exists, this function returns {:error, {:already_started, pid}} with the PID of that process.

If the init/1 callback fails with reason, this function returns {:error, reason}. Otherwise, if it returns {:stop, reason} or :ignore, the process is terminated and this function returns {:error, reason} or :ignore, respectively.

update(server, itinerary, pref_transport)

Updates the state of the UQC FSM.

Parameters

pid

Return values

Returns {:next_state, next_state, nil}.

update_bus(list)

Checks list if preferred mode of transport is bus.

update_flight(list)

Checks list if preferred mode of transport is flight.

update_train(list)

Checks list if preferred mode of transport is train.