TransportScheduler v0.1.0 InputParser

Module to store data from input files in the defined station structure

Functions

get_city_code(pid, city)

Return station city code

get_local_variables(pid, code)

Return map of station local variables

get_other_means(pid, code)

Return list of other means from station

get_schedule(pid, code)

Return schedule list of vehicles from station to destination

get_schedules(pid)

Return schedule list of vehicles from station

get_station_map(pid)

Return map of station city and code

get_station_struct(pid, city)

Return station struct

obtain_loc_var_map()

obtain_other_means()

obtain_schedules()

obtain_stations()

start_link()

Start input parser process

stop(pid)

Stop input parser process