TransportScheduler v0.1.0 Station

Module to create a transit station node. The station process pid can be used to retrieve station struct values such as local variables and schedule. It can also be used to handle local variable updates.

Uses GenStateMachine.

Functions

check_dest(dst, itinerary)

Checks if neighbouring station is valid, ie, that it does not result in a looping itinerary. The potential new destination to be added to the itinerary path is compared with all stations already in the itinerary upto this point to decide whether a loop is being formed.

Parameters

dst_station_pid

itinerary - in the form of a map %{vehicleID, src_station, dst_station, dept_time, arrival_time, mode_of_transport}

Return values

Returns true or false.

check_neighbours(schedule, other_means, time, itinerary)

Lists all valid neighbouring stations given a station and arrival time at that station. This takes care of overnight journey arrival time modifications. It also adds other means connections to the list of all possible connections to neighbouring stations. If a neighbour is valid based on check_dest/2, it is added to the list.

Parameters

schedule - in the form of a list of maps %{vehicleID, src_station, dst_station, dept_time, arrival_time, mode_of_transport}

other_means - in the form of a similar list with mode_of_transport: “Other Means”

time - time of arrival at station

itinerary - in the form of list of maps %{vehicleID, src_station, dst_station, dept_time, arrival_time, mode_of_transport}

Return values

Returns list of connections to valid neighbouring stations, each in the form of a map %{vehicleID, src_station, dst_station, dept_time, arrival_time, mode_of_transport}.

function(nc, src, itinerary, dest_schedule)

This function passes queries through the network to build the final itinerary. The connection that is required from the schedule list to reach this station from the source station is added to the itinerary path. This resulting itinerary is passed along itinerary to neighbouring stations from a given source station. If the neighbouring station happens to be the destination station required in the query, then the function passes the final itinerary directly to the Query Collector and the search is terminated.

Parameters

nc_pid

src_station_pid

itinerary - in the form of a map %{vehicleID, src_station, dst_station, dept_time, arrival_time, mode_of_transport}

dst_schedule - in the form of a list of maps %{vehicleID, src_station, dst_station, dept_time, arrival_time, mode_of_transport}

Return values

Returns {:ok}.

get_state(station)

Retrieves station struct state value given the station pid.

Parameters

station_pid

Return values

Returns {:next_state, state, vars, [{:reply, from, state}]}.

get_vars(station)

Retrieves station struct local variable values given the station pid.

Parameters

station_pid

Return values

{:next_state, state, vars, [{:reply, from, vars}]}.

receive_at_src(nc, src, itinerary)

Receives a query encoded in itinerary from NC at source station. The pid of NC and source station must be known.

Parameters

nc_pid

src_station_pid

itinerary - in the form of a map %{vehicleID, src_station, dst_station, dept_time, arrival_time, mode_of_transport}

Return values

Returns {:ok}.

send_to_stn(nc, src, dst, itinerary)

Sends query from a source station to the neighbouring station and append to itinerary built upto this point, ie, upto source station. The pid of NC, source station, and destination station must be known.

Parameters

nc_pid

src_station_pid

dst_station_pid

itinerary - in the form of a map %{vehicleID, src_station, dst_station, dept_time, arrival_time, mode_of_transport}

Return values

Returns {:ok}

start_link()

Starts a GenStateMachine Station process linked to the current process.

This is often used to start the GenStateMachine as part of a supervision tree.

Once the server is started, the init/1 function of the given module is called with args as its arguments to initialize the server.

Parameters

module

args

Return values

If the server is successfully created and initialized, this function returns {:ok, pid}, where pid is the PID of the server. If a process with the specified server name already exists, this function returns {:error, {:already_started, pid}} with the PID of that process.

If the init/1 callback fails with reason, this function returns {:error, reason}. Otherwise, if it returns {:stop, reason} or :ignore, the process is terminated and this function returns {:error, reason} or :ignore, respectively.

update(station, new_vars)

Updates station struct values by replacing with new struct composed of the new values for the station, and update the state of the station FSM.

Parameters

station_pid

station_struct

Return values

Returns {:next_state, next_state, new_vars}.